// src/service/spell/model/get-spell-response-data-model.ts
export default interface GetSpellResponseDataModel {
  SummonerBarrier: {
    key: string
    name: string
    description: string
    image: {
      full: string
    }
  }
  SummonerBoost: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerDot: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerExhaust: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerFlash: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerHaste: {
    description: string

    key: string
    name: string
    image: {
      full: string
    }
  }
  SummonerHeal: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerMana: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerPoroRecall: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerPoroThrow: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerSmite: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerSnowURFSnowball_Mark: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerSnowball: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  SummonerTeleport: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  Summoner_UltBookPlaceholder: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
  Summoner_UltBookSmitePlaceholder: {
    key: string
    name: string
    description: string

    image: {
      full: string
    }
  }
}